requirements.txt
asgiref>=3.2.10
autopep8==1.5
cachetools==4.0.0
certifi==2022.12.7
chardet==3.0.4
Django==3.1.14
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-js-asset==1.2.2
djangorestframework==3.11.2
djangorestframework-simplejwt==4.4.0
idna==2.9
Pillow>=7.1.0
protobuf==3.18.3
pycodestyle==2.5.0
PyJWT==2.4.0
python-dotenv==0.13.0
pytz==2019.3
razorpay==1.2.0
requests
six==1.14.0
sqlparse==0.3.1
stripe==2.48.0
urllib3==1.26.5